Add files via upload
kv file contains the layout of the app, the getQR.py file contain the functions allowing the app to retrieve the qr codes from izzly (can be used separately), and it's in this file that you must put in your email and passworld for it to work. the main.py contains the function called at the click of the button and the necessary imports for the kivy app to be launched. image.png is the image of the qr code that is registered each time the button is clicked.